+++ kedro_sagemaker/cli_functions.py
@@ -1,4 +1,6 @@
+import importlib
 import json
+import logging
 from contextlib import contextmanager
 from pathlib import Path
 from typing import Callable, Iterator, Optional, Tuple
@@ -6,6 +8,7 @@ from typing import Callable, Iterator, Optional, Tuple
 import click
 from sagemaker.workflow.pipeline import Pipeline as SageMakerPipeline
 
+from kedro_sagemaker.constants import MLFLOW_TAG_EXECUTION_ARN
 from kedro_sagemaker.generator import KedroSageMakerGenerator
 from kedro_sagemaker.utils import (
     CliContext,
@@ -13,6 +16,8 @@ from kedro_sagemaker.utils import (
     docker_build,
     docker_push,
 )
+
+logger = logging.getLogger()
 
 
 def parse_extra_params(params, silent=False):
@@ -94,3 +99,26 @@ def write_file_and_confirm_overwrite(
         filepath.write_text(contents)
     elif on_denied_overwrite:
         on_denied_overwrite(filepath)
+
+
+def lookup_mlflow_run_id(context, sagemaker_execution_arn: str):
+    import mlflow
+    from kedro_mlflow.config.kedro_mlflow_config import KedroMlflowConfig
+
+    mlflow_conf: KedroMlflowConfig = context.mlflow
+    mlflow_runs = mlflow.search_runs(
+        experiment_names=[mlflow_conf.tracking.experiment.name],
+        filter_string=f'tags.`{MLFLOW_TAG_EXECUTION_ARN}` = "{sagemaker_execution_arn}"',
+        max_results=1,
+        output_format="list",
+    )
+    importlib.reload(mlflow.tracking.request_header.registry)
+
+    if len(mlflow_runs) == 0:
+        logger.warning(
+            "Unable to find parent mlflow run id for the current execution (%s)",
+            sagemaker_execution_arn,
+        )
+        return mlflow.tracking._RUN_ID_ENV_VAR, None
+
+    return mlflow.tracking._RUN_ID_ENV_VAR, mlflow_runs[0].info.run_id