+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-02-08
+
+-   Support for Mlflow with shared run across pipeline steps
+-   Fixed ability to overwrite docker image in `kedro sagemaker run`
+
 ## [0.1.1] - 2022-12-30
 
 -   Pass missing environment to the internal entrypoint
@@ -15,7 +20,9 @@
 
 -   Project seed prepared
 
-[Unreleased]: https://github.com/getindata/kedro-sagemaker/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/getindata/kedro-sagemaker/compare/0.2.0...HEAD
+
+[0.2.0]: https://github.com/getindata/kedro-sagemaker/compare/0.1.1...0.2.0
 
 [0.1.1]: https://github.com/getindata/kedro-sagemaker/compare/0.1.0...0.1.1